# data-service/backend/api/data/v1/openapi.yaml
openapi: 3.0.4
info:
  title: Tidewise Data Service API
  version: 1.0.0
  description: >-
    Data Service 内部 HTTP REST + JSON 合同。消费方拥有小型手写 typed client，
    并以本文件的 operationId、schema drift anchor 和静态合同测试防止漂移。
...
servers:
  - url: /
    description: 由运行环境注入 Data Service origin；合同只冻结路径。
security:
  - ServiceBearer: []
...
paths:
...
  /api/data/v1/event-semantics/eligible-events:
    get:
      tags: [Event Semantics]
      operationId: listEligibleEventSemanticEvents
      x-client-drift-anchor: data.v1.listEligibleEventSemanticEvents
      x-required-service-scope: data.event-semantics.read
      x-retry-policy: safe-get
      x-timeout-budget-ms: 5000
      summary: 列出可创建首次语义分析 Work Item 的 Event
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          description: 由上一页返回的 opaque cursor。
          schema:
            type: string
            maxLength: 1024
        - name: pagination
          in: query
          required: false
          description: 新消费者显式设置为 cursor 后，响应才返回 next_cursor；省略时保持旧版 only-events 响应。
          schema:
            type: string
            enum: [cursor]
      responses:
        '200':
          description: Eligible Event 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EligibleEventSemanticEventsEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/data/v1/event-semantics/context-leases:
    post:
      tags: [Event Semantics]
      operationId: createEventSemanticContextLease
      x-client-drift-anchor: data.v1.createEventSemanticContextLease
      x-required-service-scope: data.event-semantics.write
      x-retry-policy: no-automatic-retry
      x-timeout-budget-ms: 5000
      summary: 为 AgentRun 已领取的 Work Item 创建数据侧 Context Lease
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventSemanticContextLeaseRequest'
      responses:
        '201':
          description: Context Lease 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSemanticContextLeaseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/data/v1/event-semantics/context-leases/{context_lease_id}/context:
    get:
      tags: [Event Semantics]
      operationId: getEventSemanticContext
      x-client-drift-anchor: data.v1.getEventSemanticContext
      x-required-service-scope: data.event-semantics.read
      x-retry-policy: safe-get
      x-timeout-budget-ms: 5000
      summary: 获取 Context Lease 对应 Event 与版本化本体上下文
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: context_lease_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Event Semantic Context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSemanticContextEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/EventSemanticContextDrift'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/data/v1/event-semantics/submissions:
    post:
      tags: [Event Semantics]
      operationId: createEventSemanticSubmission
      x-client-drift-anchor: data.v1.createEventSemanticSubmission
      x-required-service-scope: data.event-semantics.write
      x-retry-policy: idempotent-with-agent-execution-id-and-canonical-payload
      x-timeout-budget-ms: 15000
      x-atomicity: whole-submission-candidate-snapshot
      summary: 持久化一个 Event 的候选快照并执行确定性预检
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventSemanticSubmissionRequest'
      responses:
        '200':
          description: 相同 canonical payload 的安全重放
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSemanticSubmissionEnvelope'
        '201':
          description: Submission 与候选快照已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSemanticSubmissionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/data/v1/event-semantics/submissions/{submission_id}/reviews:
    post:
      tags: [Event Semantics]
      operationId: submitEventSemanticReview
      x-client-drift-anchor: data.v1.submitEventSemanticReview
      x-required-service-scope: data.event-semantics.write
      x-retry-policy: idempotent-with-reviewer-execution-key-and-canonical-payload
      x-timeout-budget-ms: 15000
      x-atomicity: whole-review-snapshot
      summary: 提交独立 Reviewer 或 Adjudicator 结构化结果
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: submission_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventSemanticReviewRequest'
      responses:
        '200':
          description: Review 已保存并由 Data 完成状态裁决
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSemanticSubmissionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/data/v1/events/{event_id}/semantics:
    get:
      tags: [Event Semantics]
      operationId: getEventSemantics
      x-client-drift-anchor: data.v1.getEventSemantics
      x-required-service-scope: data.event-semantics.read
      x-retry-policy: safe-get
      x-timeout-budget-ms: 5000
      summary: 读取 Event 的完整语义结果与审计状态
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: event_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Event Semantics 结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSemanticsEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    ServiceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque service identity token
      description: >-
        环境 secret 注入的 service identity。服务端从认证 principal 派生 caller identity
        与 scope；request body 不得声明或覆盖 caller identity。
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: 可选 caller request id；缺失时服务端生成。不得作为业务幂等 identity。
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  headers:
    RequestID:
      description: 本次 transport attempt 的 request id；重放时可以与首次请求不同。
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: 请求 DTO、query 或边界不合法
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: service bearer identity 缺失或无效
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 已认证 principal 不具有 operation 所需 scope
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: caller scope 内不存在目标资源
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: 幂等合同或自然身份对应内容冲突；零 mutation
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    EventSemanticContextDrift:
      description: Context Lease、分页请求身份或所选正式路径已漂移；调用方应以新 lease 重试
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EventSemanticContextDriftErrorEnvelope'
    PayloadTooLarge:
      description: JSON request body 超过 1048576 bytes
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnprocessableEntity:
      description: 已解析 DTO 未通过 whole-batch 业务 validation；零 mutation
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalError:
      description: Data Service 内部错误
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServiceUnavailable:
      description: Event Semantics 依赖暂时不可用
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    UUID:
      type: string
      format: uuid
      description: RFC 4122 UUID string。
...
    UTCTimestamp:
      type: string
      format: date-time
      description: UTC RFC3339 timestamp；response 必须使用 UTC offset Z。
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
          description: 稳定 machine-readable error code。
        message:
          type: string
          minLength: 1
          maxLength: 500
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      additionalProperties: false
      x-client-drift-anchor: data.v1.schema.ErrorEnvelope
      required: [error, request_id]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string
          minLength: 1
          maxLength: 128
...
    EventSemanticContextLeaseRequest:
      type: object
      additionalProperties: false
      required: [event_id, agent_execution_id, worker_id, lease_seconds]
      properties:
        event_id:
          $ref: '#/components/schemas/UUID'
        supersedes_submission_id:
          $ref: '#/components/schemas/UUID'
        agent_execution_id:
          type: string
          minLength: 1
          maxLength: 200
        worker_id:
          type: string
          minLength: 1
          maxLength: 200
        lease_seconds:
          type: integer
          minimum: 60
          maximum: 900
    EventSemanticContextLease:
      type: object
      additionalProperties: false
      required: [context_lease_id, event_id, status, lease_expires_at]
      properties:
        context_lease_id:
          $ref: '#/components/schemas/UUID'
        event_id:
          $ref: '#/components/schemas/UUID'
        supersedes_submission_id:
          $ref: '#/components/schemas/UUID'
        status:
          type: string
          enum: [active]
        lease_expires_at:
          $ref: '#/components/schemas/UTCTimestamp'
    EventSemanticContextLeaseEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, result]
      properties:
        request_id:
          type: string
        result:
          $ref: '#/components/schemas/EventSemanticContextLease'
    EligibleEventSemanticEvent:
      type: object
      additionalProperties: false
      required: [event_id]
      properties:
        event_id:
          $ref: '#/components/schemas/UUID'
    EligibleEventSemanticEvents:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/EligibleEventSemanticEvent'
        next_cursor:
          type: string
          description: 下一页 opaque cursor；缺失表示已到末页。
    EligibleEventSemanticEventsEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, result]
      properties:
        request_id:
          type: string
        result:
          $ref: '#/components/schemas/EligibleEventSemanticEvents'
    EventSemanticEvent:
      type: object
      additionalProperties: false
      required: [id, title, summary, occurred_at, event_status, fact_status]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
        summary:
          type: string
        occurred_at:
          allOf:
            - $ref: '#/components/schemas/UTCTimestamp'
          nullable: true
        event_status:
          type: string
          enum: [confirmed]
        fact_status:
          type: string
          enum: [verified]
    EventSemanticEvidence:
      type: object
      additionalProperties: false
      required:
        - evidence_id
        - evidence_hash
        - evidence_statement
        - source_level
        - relation
        - supports_fields
        - raw_document_id
        - source_name
        - source_type
        - title
        - first_seen_at
        - knowledge_available_at
        - accepted_at
        - statement_source
      properties:
        evidence_id:
          $ref: '#/components/schemas/UUID'
        evidence_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        evidence_statement:
          type: string
          minLength: 1
        source_level:
          type: string
          enum: [primary, secondary]
        relation:
          type: string
          enum: [supports, contradicts, context]
        supports_fields:
          type: array
          items:
            type: string
        raw_document_id:
          $ref: '#/components/schemas/UUID'
        source_name:
          type: string
          minLength: 1
        source_type:
          type: string
          minLength: 1
        source_url:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
        published_at:
          allOf:
            - $ref: '#/components/schemas/UTCTimestamp'
          nullable: true
        first_seen_at:
          $ref: '#/components/schemas/UTCTimestamp'
        knowledge_available_at:
          $ref: '#/components/schemas/UTCTimestamp'
        accepted_at:
          $ref: '#/components/schemas/UTCTimestamp'
        statement_source:
          type: string
          description: Event Fact 明确提供的原始声明主体；空字符串表示正式事实尚未提供，不能用媒体来源名补齐。
...
    EventSemanticVariableDefinition:
      type: object
      additionalProperties: false
      required: [key, version, name_zh, name_en, domain, business_definition, value_type, status, allowed_directions, allowed_units, applicable_entity_types]
      properties:
        key:
          type: string
        version:
          type: integer
          minimum: 1
        name_zh:
          type: string
        name_en:
          type: string
        domain:
          type: string
        business_definition:
          type: string
        value_type:
          type: string
        status:
          type: string
        allowed_directions:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticDirection'
        allowed_units:
          type: array
          items:
            type: string
        applicable_entity_types:
          type: array
          items:
            type: string
...
    EventSemanticContext:
      type: object
      additionalProperties: false
      required: [context_lease_id, agent_execution_id, worker_id, lease_expires_at, manifest_contract_version, context_fingerprint, event_fingerprint, evidence_fingerprint, ontology_version, acceptance_policy_version, event, evidence, variable_definitions, assertion_modalities, measurement_contract]
      properties:
        context_lease_id:
          $ref: '#/components/schemas/UUID'
        agent_execution_id:
          type: string
        worker_id:
          type: string
        lease_expires_at:
          $ref: '#/components/schemas/UTCTimestamp'
        manifest_contract_version:
          type: string
          enum: [event-semantic-context-manifest.v4]
        context_fingerprint:
          type: string
          pattern: '^[0-9a-f]{64}$'
        event_fingerprint:
          type: string
          pattern: '^[0-9a-f]{64}$'
        evidence_fingerprint:
          type: string
          pattern: '^[0-9a-f]{64}$'
        ontology_version:
          type: string
        acceptance_policy_version:
          type: string
        event:
          $ref: '#/components/schemas/EventSemanticEvent'
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticEvidence'
        variable_definitions:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticVariableDefinition'
        assertion_modalities:
          type: array
          items:
            type: string
        measurement_contract:
          $ref: '#/components/schemas/EventSemanticMeasurementContract'
    EventSemanticMeasurementContract:
      type: object
      additionalProperties: false
      required: [representation, max_items_per_signal, max_text_characters, requires_evidence_ids, numeric_validation]
      properties:
        representation: { type: string, enum: [evidence_grounded_narrative] }
        max_items_per_signal: { type: integer, minimum: 1, maximum: 20 }
        max_text_characters: { type: integer, minimum: 1, maximum: 10000 }
        requires_evidence_ids: { type: boolean, enum: [true] }
        numeric_validation: { type: boolean, enum: [false] }
...
    EventSemanticContextEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, result]
      properties:
        request_id:
          type: string
        result:
          $ref: '#/components/schemas/EventSemanticContext'
    EventSemanticContextDriftErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
          enum: [EVENT_SEMANTIC_CONTEXT_DRIFT]
        message:
          type: string
          minLength: 1
          maxLength: 500
        details:
          type: object
          additionalProperties: true
    EventSemanticContextDriftErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          $ref: '#/components/schemas/EventSemanticContextDriftErrorDetail'
        request_id:
          type: string
          minLength: 1
          maxLength: 128
    EventSemanticDirection:
      type: string
      enum: [increase, decrease, unchanged, mixed, uncertain]
    EventSemanticMeasurement:
      type: object
      additionalProperties: false
      required: [measurement_text, evidence_ids]
      properties:
        measurement_text:
          type: string
          minLength: 1
          maxLength: 2000
        evidence_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/UUID'
    EventSemanticEntityLinkCandidate:
      type: object
      additionalProperties: false
      required: [candidate_key, mention, entity_id, entity_role, evidence_ids, resolution_method]
      properties:
        candidate_key:
          type: string
        mention:
          type: string
        entity_id:
          $ref: '#/components/schemas/UUID'
        projected_entity_type:
          type: string
          minLength: 1
        entity_role:
          type: string
        evidence_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/UUID'
        resolution_method:
          type: string
          enum: [qdrant_exact, qdrant_vector]
        resolution_confidence:
          type: string
          pattern: '^(0(\.\d{1,5})?|1(\.0{1,5})?)$'
    EventSemanticV3EntityLinkCandidate:
      allOf:
        - $ref: '#/components/schemas/EventSemanticEntityLinkCandidate'
        - type: object
          required: [projected_entity_type]
    EventSemanticVariableSignalCandidate:
      type: object
      additionalProperties: false
      required: [candidate_key, subject_link_key, variable_key, variable_version, direction, assertion_modality, evidence_ids, measurements]
      properties:
        candidate_key:
          type: string
        subject_link_key:
          type: string
        variable_key:
          type: string
        variable_version:
          type: integer
          minimum: 1
        direction:
          $ref: '#/components/schemas/EventSemanticDirection'
        assertion_modality:
          type: string
          enum: [actual, stated_intent, source_forecast]
        evidence_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/UUID'
        measurements:
          type: array
          maxItems: 8
          items:
            $ref: '#/components/schemas/EventSemanticMeasurement'
        statement_at:
          $ref: '#/components/schemas/UTCTimestamp'
        valid_from:
          $ref: '#/components/schemas/UTCTimestamp'
        valid_until:
          $ref: '#/components/schemas/UTCTimestamp'
        forecast_period_start:
          $ref: '#/components/schemas/UTCTimestamp'
        forecast_period_end:
          $ref: '#/components/schemas/UTCTimestamp'
        extraction_confidence:
          type: string
          pattern: '^(0(\.\d{1,5})?|1(\.0{1,5})?)$'
    EventSemanticSubmissionRequest:
      type: object
      additionalProperties: false
      required: [context_lease_id, event_id, agent_execution_id, agent_key, agent_version, generator_prompt_hash, generator_model, reviewer_prompt_hash, reviewer_model, adjudicator_prompt_hash, adjudicator_model, ontology_version, acceptance_policy_version, entity_links, variable_signals]
      properties:
        context_lease_id:
          $ref: '#/components/schemas/UUID'
        event_id:
          $ref: '#/components/schemas/UUID'
        agent_execution_id:
          type: string
        agent_key:
          type: string
          enum: [event-semantic-enricher]
        agent_version:
          type: string
          enum: [event-semantic-enricher.v3]
        supersedes_submission_id:
          $ref: '#/components/schemas/UUID'
        generator_prompt_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        generator_model:
          type: string
        reviewer_prompt_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        reviewer_model:
          type: string
        adjudicator_prompt_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        adjudicator_model:
          type: string
        ontology_version:
          type: string
        acceptance_policy_version:
          type: string
        entity_links:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticV3EntityLinkCandidate'
        variable_signals:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticVariableSignalCandidate'
...
    EventSemanticSubmissionResult:
      type: object
      additionalProperties: false
      required: [submission_id, event_id, status, canonical_payload_hash, replayed, entity_links, variable_signals]
      properties:
        submission_id:
          $ref: '#/components/schemas/UUID'
        event_id:
          $ref: '#/components/schemas/UUID'
        status:
          type: string
          enum: [pending_review, needs_reanalysis, quarantined, accepted, rejected, superseded]
        canonical_payload_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        replayed:
          type: boolean
        entity_links:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticCandidateDecision'
        variable_signals:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticCandidateDecision'
        reviewer_work_package:
          $ref: '#/components/schemas/EventSemanticReviewerWorkPackage'
        audit_work_package:
          $ref: '#/components/schemas/EventSemanticReviewerWorkPackage'
        context_lease_id:
          $ref: '#/components/schemas/UUID'
        agent_execution_id:
          type: string
        agent_key:
          type: string
          enum: [event-semantic-enricher]
        agent_version:
          type: string
          enum: [event-semantic-enricher.v2, event-semantic-enricher.v3]
        supersedes_submission_id:
          $ref: '#/components/schemas/UUID'
        generator_prompt_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        generator_model:
          type: string
        reviewer_prompt_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        reviewer_model:
          type: string
        adjudicator_prompt_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        adjudicator_model:
          type: string
        ontology_version:
          type: string
        acceptance_policy_version:
          type: string
        candidate_snapshot:
          type: object
          additionalProperties: true
        review_snapshots:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticReviewSnapshot'
        created_at:
          type: string
          format: date-time
        finalized_at:
          type: string
          format: date-time
    EventSemanticSubmissionEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, result]
      properties:
        request_id:
          type: string
        result:
          $ref: '#/components/schemas/EventSemanticSubmissionResult'
    EventSemanticReviewItem:
      type: object
      additionalProperties: false
      required: [candidate_type, candidate_key, decision, reason_codes, evidence_ids]
      properties:
        candidate_type:
          type: string
          enum: [entity_link, variable_signal]
        candidate_key:
          type: string
        decision:
          type: string
          enum: [pass, fail, indeterminate]
        reason_codes:
          type: array
          items:
            type: string
        evidence_ids:
          type: array
          items:
            $ref: '#/components/schemas/UUID'
    EventSemanticReviewRequest:
      type: object
      additionalProperties: false
      required: [reviewer_execution_key, prompt_hash, model, items]
      properties:
        reviewer_execution_key:
          type: string
        prompt_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        model:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/EventSemanticReviewItem'
    EventSemanticsResult:
      type: object
      additionalProperties: false
      required: [event_id, submissions]
      properties:
        event_id:
          $ref: '#/components/schemas/UUID'
        submissions:
          type: array
          items:
            $ref: '#/components/schemas/EventSemanticSubmissionResult'
    EventSemanticsEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, result]
      properties:
        request_id:
          type: string
        result:
          $ref: '#/components/schemas/EventSemanticsResult'